Move enumeration module into data module
